+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.4 (unreleased)
 
+* Added setting `ENVELOPE_USER_CHANNEL_SEND_SUBSCRIBE` to send a subscribe message to consumer rather than just adding
+  the client to the users channel.
 * Loading RQ-job via message class for deferred jobs, to make the code easier to follow + overrides simpler.
 
 ## 1.0.3 (2024-03-13)